Remove explicit incrementing and decrementing of domain ref counts in arrays.
Remove the routines chpl_incRefCountsForDomainsInArrayEltTypes() and
chpl_decRefCountsForDomainsInArrayEltTypes() and calls to them.  This was
previous workaround for the fact that "nude" domains are referenced in array
implementations and the implementations did not track the reference counts
directly.

This workaround is no longer necessary, since as a special case AMM now performs
reference count updates on "nude" record-wrapped types.  This workaround was
also incomplete, because it deals only with domain reference counts and not array and
distribution reference counts.

The special handling in AMM can also be considered to be a workaround.  The
"correct" implementation would rely on simply incrementing the reference count
when a new reference is creating and decrementing it when a reference is
destroyed.  A reference in this context means a variable or field containing a
(class) variable that is a copy of the "nude" reference-counted type (domain,
array or distribution).

Since the _array.initialize() routine contained only the call to
chpl_incRefCountsForDomainsInArrayEltTypes(), it could be removed entirely.