Support complex.re and complex.im as reference functions using C99 complex
In the setter side of complex.re and complex.im, call the complex_get_real
and complex_get_imag primitives.  Let the primitives exist until codegen,
where they are generated as calls to functions in the runtime.

The new runtime functions take a pointer to complex and return a pointer to
the real or imaginary component.  They treat the complex number as a
two-element array of reals containing [real, imag].

The inlineFunctions pass was changing the reference argument in these
primitives into non-references, so modified it to leave them as references.

Updated loopInvariantCodeMotion to consider the argument to
complex_get_real/imag and the thing it returns to be aliases.