Move _defaultOf closer to involved types
To resolve some function resolution order problems, I've moved the
implementation of _defaultOf from Types.chpl to ChapelBase.chpl for most
things. Moved _defaultOf for memory_order to MemConsistency.chpl;
added _defaultOf for c_ptr and c_void_ptr to SysBasic.chpl instead
of relying on a catch-all case.

Removes the catch-all case for _defaultOf which used to be
  inline proc _defaultOf(type t) {
    select t {
      when memory_order {
        return _defaultOfMemoryOrder();
      }
      otherwise {
        return nil:t;
      }
    }
  }

and is apparently no longer needed.