Fix the two immediate errors testing with resolveUnresolvedSymExpr revealed:
For some reason, I was storing the number of symbols found as a boolean, causing
incorrect behavior.  Must've been a typo/brain freeze.  Additionally, switching
to using STL vectors for storage instead of our Vec storage containers meant
that we no longer had exclusive add for insertion into the container (so we
were getting double insertions of the same Symbol).