Fix null byte in output bug.
Added lots of comments to the source code about the related
behavior. The bug was because snprintf(buf, n, fmt) returns
the number of bytes needed for the output not including a '\0',
but it always null terminates. So we need to account for the
null terminator to know if the conversion was successful.