Update the future no-use-record
Michael's movement of _defaultOf around in the standard modules yesterday
caused this future to fail in a different (though similar) manner than it did
before.  In addition to updating the .bad file, I added some of the strategies
for resolving the issue that Brad brought up in our discussion via email.