Fix foralls over arrays to use standalone iterator.
Prior to this change, leader/follower would be invoked
instead of the standalone iterator in a forall over an array
when there is an outer variable in the loop body, e.g.

```
    var b ...;
    forall a in A do
      a = b;
```

DISCUSSION

We fix this by adding another case of proc _toStandalone().

That added case works due to using _build_tuple_always_allow_ref()
in propagateExtraLeaderArgs() / implementForallIntents.cpp

Which, in turn, is enabled by PR #2078.

TEST CHANGES

This test:

  test/arrays/deitz/parallelism/test_whole_array_assignment_of_elemental_value_is_parallel.chpl

now uses standalone instead of leader/follower, so I updated its .good.

While there:

* I added a printout in DefaultRectangular for standalone to mirror
the one that was there for the follower. That way, the above test
can ensure that the standalone iterator is indeed invoked.

* I added two more foralls to the above test to ensure that
the standalone iterator is used there as well,
including the motivating example at the beginning of this message.

* BTW if I compile the NEW version of that test with the OLD version
of Chapel, it reports DR domain follower for the first forall
(without the 'with' clause) and NOT for the second forall
(with "with (ref r)"). I imagine both foralls should behave the same.
I did not investigate.

* I fixed a bug in PREDIFF for this test.
Since PREDIFF in stream/ subdirectory was its exact copy,
I replaced it with a symlink.