Resolve main iterator inlining bug on string-as-rec branch.
Reference counts were going below zero leading to premature deletion when
inlined iterators were enabled for certain test cases.

Wholly different code is used to implement inlined vs. non-inlined iterators.
Integer indices might be returned from the non-inlined version, while domains
or arrays are returned from the inlined version.

The latter version was failing to insert an autocopy before the return in the
"tuple.this" function, expanded to return a wrapped array type.  I traced the
problem down to the PRIM_GET_SVEC_MEMBER_VALUE call in the implementation of
tuple.this().  Because this primitive returns an element from an array, the
returned value is unowned.  I had to add that primitive to the explicit list
that claims the result is unowned.  With that change the needed autocopies are
inserted.

I updated or removed the compopts files corresponding to the tests that were
failing that now work.  There are two test cases that were not cured by this
fix.  Further investigation is required.