Fix tcmalloc when compiled with clang 3.6 and used from C++
tcmalloc.h uses struct mallinfo (it returns it from tc_mallinfo.
That gives an error like this:
.../include/gperftools/tcmalloc.h:101:38: error:
  'tc_mallinfo' has C-linkage specified, but returns incomplete type 'struct mallinfo' which could
        be incompatible with C [-Werror,-Wreturn-type-c-linkage]
          PERFTOOLS_DLL_DECL struct mallinfo tc_mallinfo(void) __THROW;
                                               ^
struct mallinfo is not a complete type because malloc.h is not included in the
user-facing tcmalloc.h. It looks like tcmalloc.h includes malloc.h but it doesn't
because it doesn't first include config.h - instead it relies on autoconf applied
to tcmalloc.h.in - and so the variables like HAVE_MALLOC_H are always undefined.

This patch solves the problem by creating autoconf variables for the user-facing
tcmalloc.h to use to include the header file defining struct mallinfo.