Add --break-on-resolve-id compiler option
This enables faster debugging of function resolution.
I find that I'm frequently trying to understand why
a particular call didn't resolve. To do so, I have
been using --explain-call-id and setting a breakpoint
in the output there. Since I was doing that often,
I created a patch to provide that functionality more
directly.

--break-on-resolve id will break in two places:
 - in resolveNormalCall
 - when used with --explain-call-id and --explain-verbose,
   before every candidate is filtered