Removing changes for static methods that were no longer necessary
As I went from an INTENT_TYPE approach to more of the build.cpp-based
transform to the 'this' argument that we apply to other 'type'
arguments, most of the changes I made in function resolution were
unnecessary.  The only one remaining is to nullify the 'this'
argument if the 'this' formal argument is removed from the AST.

I also removed the .future for Thomas's original typeProcs
future since it is now working.