Several changes to the implementation (all good going forward, me thinks)
First, I allowed function names to override method names in the symbol table if
the methods were found first.  As part of this change, I also tightened the name
conflict check to reject programs where in the same scope a function and another
symbol (not a method or other function) were defined with the same name.

Then, I removed all checks against the old lookup implementation, as I believe
mine is now superior.

With the removal of all these checks, the old lookup implementation was no
longer used, so I removed its code from the file.