Clean up comments and move common code after symbol found in symbolTable
My comments were a little out of date and sparse.  There was a common check
after finding a symbol in the symbol table for a module or other symbol to make
sure that if the found symbol was a method, it was visible in this scope in its
pure named form.  I moved that check into the symbol table helper.