Round 2 of working around knc module bug
We're having issues with knc testing on kay-esl. See the thread titled "modules
on kay-esl" It's a pretty specific combination of running on an esl, using the
craype-intel-knc module, and passing -V to the qsub reservation to preserve the
environment.

We're trying to figure out a long term solution to what we believe is a bug in
the knc module. In the meantime, unloading the knc module before doing our qsub
reservation works around the issue. Note that this patch is careful to make
sure self.knc will still return true even after we unload the knc module since
chpl_launchcmd relies on self.knc to add extra flags to the qsub reservation.