Allow extern records to be renamed like extern functions
We currently allow extern functions to have their symbol renamed at the
C level. This patch extends that functionality to classes, records, and
unions. This is useful when needing to work with un-typedef'd structs.

An of this example is the stat struct which does not have a typedef and
must be referred to using the 'struct' qualifier.

    extern "struct stat" record stat_chpl {
      ...
    }

I've also allowed both functions and aggregate types to be renamed using
a plain identifier or a string. The string syntax is necessary for the
case above. This makes both of the definitions below equivalent:

    extern bar proc foo() {...}
    extern "bar" proc foo() {...}