Add -valgrindexe option to paratest.
Before, if you typed -valgrindexe as an option to paratest, it would parse the
flag as -valgrind and run valgrind on the compiler as well as the test program.
This is not what the user expects.

I modified the server and client scripts to honor the "exe" suffix if present,
and in that case run valgrind on only the test program.