First working draft of static primary methods
This approach is a little sloppy:  I attached a flag to the function
overall so that, later, when its 'this' argument is being created
and added to it, it's marked appropriately.  It seems that there are
already cases that use 'this' 'ref' intents for primary methods (?!)
so it's worth looking into how they handle this.