Add a top level make rule for the parser
Adds a top level rule for the parser to make development easier. I've
also changed the build rules for the parser to be a bit smarter. Builds
should now properly detect changes to chapel.ypp and chapel.lex and
rebuild when necessary rather than every time you make the parser.