Remove extraneous #includes of scopeResolve.h
Both externCResolve and clangUtil had a #include for scopeResolve.h, but no
uses of the functions defined in that header file.  This seemed like an
oversight.  In the interest of moving a function definition to scopeResolve.h
and not having it bleed warrantlessly, I'm investigating the effects of
removing these unnecessary calls.