Undo .good update to studies/graph500/v2/main and resolve baseline failure
Honestly, this was just laziness on my part.  scalableDataGenerator had a
function defined whose arguments were shadowed by the globals in one of the
modules it used.  I had accepted this shadowing as "well, I don't *really*
know what the author intended here" and updated the .good to include the new
error messages for when this situation arises.  However, the number of times
this error message would occur varied between standard runs and baseline runs.
Really, I just needed to assume the author intended to refer to the function
arguments, because why else would they have provided them?  This commit alters
the name of the arguments so that conflicts no longer occur, reverting the
.good for the test to its original state.  Sorry for the churn.