Eliminate transitive traversal of used private modules in scopeResolve
Unfortunately, this does not prevent functions in private modules from being
found when they are used by their parent module and their parent module is
used.  That will require a function resolution fix, which I'm waiting for an
okay from Mike to perform.  All but one of the new tests pass.

Also updated an error message.