Fix 32 bit portability issue for getFileSize()
Our Linux 32 testing failed on the tests for getFileSize because we had
hard-coded the definition of off_t at the Chapel level to be 64 bits when it
would be 32 bits on 32 bit platforms.  Unfortunately, there is not a good way
to vary the size of the Chapel definition of off_t similar to our other C
types - off_t is not as standard as the others and its size variations are not
as easily visible.  The portable solution recommended is to expect the largest
size available.  Which isn't future-proof, but so it goes.  With this in mind,
instead of having the runtime support functions give an off_t, I will have them
return an int64_t.