Add description of static methods / 'type' this intents; wordsmith
I added a description of how to declare static methods to the
section describing this intents and took a wordsmithing pass
on these paragraphs while here, which included splitting the
this intents into those that are constraints on the receiver
(param and this) vs. those that are more about "how to pass"
(ref and blank).

Added a test of param and type this-intents.

Added new index terms for static methods.