Name dereference temps uniformly for greppability
Added names for dereference temporaries where they were missing and modified
existing names so that most dereference temps now at least contain "derefTmp"
or "DerefTmp".  This will make them easy to identify in generated code.

Except for fundamental types and all but the simplest of record types, passing
objects by value is undesirable in the general case.

The rules change when the object is being passed across the network.  There, it
is desirable to send larger structures across when it is safe to do so.  The
Remote Value Forwarding (RVF) optimization handles this case and uses special
"RVF" markers on its dereference temporaries (on the receiving end).