Rewording description of "static methods" to avoid that term
Mike pointed out that my understanding of static methods in C++
was incomplete/incorrect due to the fact that they can be invoked
on instances as well as types -- different than what I'd
implemented or documented for Chapel.

To that end, switch to the terms "instance methods" vs. "type
methods".  While doing this, I also took a pass over text
relating to receivers and methods in general, including adding
mention of primary and secondary methods (which we seemingly
didn't refer to in the spec, as I thought we had); and mention
that secondary methods can be defined on arbitrary types.