Call cygwin's kill in timedexec since perl's doesn't always work
We noticed that the `kill` we use in timedexec wasn't always successful on
cygwin for some reason. I think this is because for some tests cygwin had some
resources that windows technically owned and regular kill didn't cut it. Now we
also call cygwin's kill "/bin/kill" and pass "-f" to it so that cygwin will use
the win32 interface to kill if it has to.

For more info see: https://cygwin.com/cygwin-ug-net/using-utils.html#kill

I tested on cygwin with release/examples/primers/fileIO. This test caused
issues after 5 trials before this change, and "worked" for 100 trials after.
"worked" here means it timed out 20 times, but we actually killed the runaway.

I protected the call with `if ("$^O" eq "cygwin")` instead of checking
CHPL_TARGET_PLATFORM since we don't have easy access to that, and timedexec is
pretty chapel agnostic