First cut at adjusting the reference level of formals to match their intents
In resolvedIntents(), when the concrete intent is ref or const ref and the
formal argument type is a value type, replace its type with its ref type.
Then, run insertReferenceTemps() and insertDerefTemps() to ensure that formal
and actual types line up.

DETAILS
###compiler/AST/expr.cpp
Added a clause to codegen for PRIM_ASSIGN to handle the case when both
arguments are of ref type.  In this case (unlike PRIM_MOVE), the value of the
RHS is copied onto the value referred to by the LHS.

###compiler/include/passes.h
###compiler/resolution/callDestructors.cpp
Add insertDerefTemps() functions.

###compiler/main/checks.cpp
Add calls to checkFormalActualTypes match in check_flattenFunctions() and
check_cullOverReferences.

###compiler/optimizations/narrowWideReferences.cpp
Work around null pointer extracted from the wideInfoMap.  Certain fields are
not widened (even though they should be).  If we check for a dependence on one
such field, we segfault.
The workaround is reasonable, because if the field is already narrow, then
narrowing of the other symbol does not depend on our ability to narrow it.  It
is narrow de facto, so we don't have to add that narrow field as a dependence.

###compiler/passes/flattenFunctions.cpp
Call insertDerefTemps after flattenNestedFunctions(), to match the ref level of
newly-added formals.

###compiler/passes/resolveIntents.cpp
Set the default intent on extern record types to const in, to match the C idea
of default intent.  (This change will be reverted.)

Adjust the reference level of each formal argument type to match its concrete
intent.  This means that formals being passed by "ref" or "const ref" will
have ref types.

Call insertReferenceTemps() and insertDerefTemps() after this to make actual
and formal argument types match.

###compiler/resolution/lowerIterators.cpp
The formal of _getIterator function is now passed by ref, so we have to remove
the reference when looking for the iterator info associated with a given
iterator class.