Add more iterator type dereferences.
Since iterator records are now passed by reference, we have to dereference the
type of the argument of _getIterator(const ref ir:_ref(_ir)) calls to get at
the defaultInitializer field wherein the original iter function is stashed.

The original iterator function symbol populates the IteratorInfo field, so this
idiom is used often to extract that information.

This drove the number of regressions down from 1547 to 1086.