Use const in intent to force default iterator arguments to be passed by value.
This change is required because the type of a field in an iterator record (and
consequently in the iterator class) corresponding to an argument to the iterator
function is copied from the signature of the iterator function.

Since records (including tuples) are now passed by ref by default, the type of
the field is a reference.  It refers to the initializer object, which is local
to the function (e.g. "these") that creates the iterator record.  Once the
iterator record is created, the local goes away, which makes that field of the
IR always invalid.  (Any access to the value referred to by that field should --
and often does -- result in a valgrind error.)

The workaround solution is to pass these arguments by "const in" so they are
forced to be copied by value in to the iterator record.