Use "in" intent for record formals in extern proc declarations
In general, when a struct is passed to a C function, it is passed by what is
equivalent to Chapel "in" intent.

Before: Formals of extern record type were passed by value in the generated C
code, even if the computed concrete intent said they should be passed by "const ref".

Now:  In this revision, we respect the computed concrete intent.  So all
variables of record type(1) are passed by pointer in the generated C code.  If
the C code expects the struct to be passed by value instead, the author of the
Chapel interface description has to say "in" explicitly.