Revise IR layout to use value types mostly.
This change was due to many test cases failing in leader-follower iterators,
because the iterable (e.g. range, domain) returned by the leader is stored in a
variable that is local to the scope of the body of the outer iterator.  That
value goes out of scope after creation of the argument bundle and before the
task executes, which can result in a reference to freed memory.  (It can also
result in a reference to stack memory which is being used for other purposes by
the time it is read.  valgrind does not seem to detect this error, but it still
leads to incorrect program behavior.)

The fix involves forcing value types to be used for the fields of an iterator
record, with the exceptions of:
  - the _this argument of an iterator method (presumed to be read/write), and
  - the value field if the return intent of the iterator is declared to be "ref".

Down to 278 regressions with --no-inline tossed.