Re-instate removed performance key and generate bogus output for it
I didn't notice that a perfkey had been removed in the previous commit
to this file, nor would I have correctly anticipated the impact (which
is that it messes up our performance graph generation).  The work to
remove a column from the history of performance data is greater than
the work to just generate a bogus value and ignore it.

Unfortunately, this doesn't save the work of having to edit the .dat
files because rows that are missing columns cause problems for the
graph generation... :(