Fix additional issues in future (beyond the reason it was filed)
This future is intended to exercise private configuration variables.  However,
there were quite a few other errors in the code, including attempting to assign
Random() to a variable instead of new Random(), multiplying a uint(64) and a
signed integer, and returning a for loop.  These issues were all independent of
private configuration variables.  Instead of spending time hunting down
whatever other bugs lived in this future, we decided to merely simplify the
future to its most basic case.