Added future: should use chain depth matter?
From .future:

Context: symbols 'x' are defined in two different modules M11 and M222.
Those modules are 'use'ed through different-depth use chains
e.g. by the main module.

Should a reference to 'x' in the main module be an "it is ambiguous" error?
Or should M11's 'x' shadow M222's 'x' because the depth of the use chain
of M222 is longer?