Initial attempt to expand private/public support
This attempts to extend the support to module level variables, type aliases,
enum declarations, and functions.  It does not attempt to extend the support
to classes/records/unions or their fields/methods.  The extension is not
complete: it is anticipated that functions will require further tinkering,
and only the most basic of testing has occurred (i.e. hello world works, the
old module support works, and a future using a configuration constant behaves
as expected).  Full testing and the inclusion of tests will occur next.

Basically, the change includes the addition of private/public to the parsing
rules for the above symbols, the extension of the isVisible method to all
Symbols instead of just ModuleSymbls, and the expansion of the checks in
scopeResolve to allow for symbols other than modules to be private.