Update tests that ensure private symbols don't hurt public symbols at same scope
They now also include outside calls to a function defined in the same scope as
the private variable that use the private variable.  Since setting the config
via the command line would change the output of these tests, the config tests
now have a separate .good file for setting the private symbols via the command
line.