Add support for task callbacks.
This adds a callback mechanism to the tasking layer, with one complete
implementation in fifo tasking.  This allows installing functions to be
called when each task starts and ends.  Installed functions can also be
uninstalled later.  Some notes:
- This is really only intended to support internal uses such as Chapel's
  own tools.  It's not intended for users!
- Currently we only support callbacks for task start and end, although
  there have already been discussions about adding others, such as when
  the threads that host tasks switch to or from them.
- Currently we allow a maximum of 10 installed callback functions for
  each different event type.
- Callback function installation and uninstallation is not thread-safe.
  If this turns out to be a problem we can discuss making it so, though
  this will necessarily make the callbacks themselves more expensive.
- Using the callbacks is a little tricky, because you have to arrange to
  include runtime/include/chpl-tasks.h in order to get the appropriate
  definitions.  See test/runtime/gbt/tasks/callbacks/tcb.precomp for an
  only mildly outrageous way to do this.

Most of what's here is actually the tests I used to confirm that things
seemed to be working properly.

The immediate user of this is the upcoming chplvis tool, but I have
wanted something like this myself for some time and expect we will find
other uses for it with little difficulty.