Fix function resolution to ignore private functions when appropriate
Previously, all the functions in a block were considered visible and so all
with a certain name were added to the list of visible functions returned by
getVisibleFunctions.  With the addition of private functions, this was no
longer the case, so instead of appending the vector per block wholesale, we
first check if they are private and not visible to the scope of the call.