Changing void* arguments to size_t* arguments
The stride and count arguments to the chpl_comm*strd() routines
have traditionally been void* arguments, but for no reason that
I can tell (they were always cast immediately to int32_t*
arguments, so it's not as though they were actually used in
any sort of polymorphic way).

With my recent conversion of these to size_t* values, the void*
nature was hiding a failure and my hope was that changing them
to size_t* would have flagged that issue earlier.  Turns out
that's not the case because the compiler casts the actuals into
void* arguments anyway, but still -- it seems nicer to avoid
the casts in the routines themselves as well.