Add warnings when private is applied incorrectly
Such as when applied to variables within a function or nested block scope.
Also adds an error message when applied to fields and methods on a class (but
I don't think it handles secondary methods yet).

Adds tests of the function local variable, nested blocks in modules and
functions, and argument cases.  No tests yet of the fields/methods case.