Fix failure in test_glob due to iterator lowering weirdness
When PRIM_TO_LEADER, PRIM_TO_FOLLOWER and PRIM_TO_STANDALONE are resolved,
postFold() inserts a call to the corresponding _toLeader, _toFollower or
_toStandalone routine.  This call may or may not contain arguments that want to
be passed by reference (according to resolveIntents).  But substitution of
arguments with fields from the iterator record cannot be performed until that
record type is created in the lowerIterators pass.

The routine cleanupLeaderFollowerIteratorCalls() performs this substitution,
but it depends on the call that needs to be fixed up remaining unchanged
between resolution and the lowerIterators pass.  Applying
insertReferenceTemps() to that call changes the form of the call, so the
cleanup routine no longer works correctly.

As a workaround, insertReferenceTemps() is prevented from changing those calls
in particular, and the check for formal/actual type mismatches is disabled for
those calls in particular until after the lowerIterators pass has completed.

Add a global symbol "iteratorsLowered" to signal the end of the lowerIterators
pass.

Add predicates to avoid adding reference temps and avoid testing whether actual
and formal argument reference levels match before iterator lowering is complete.