Fix problems in the tasking layer callback tests.
The main problem fixed here is that the "wait-for-nCallbacks" values
were wrong.  When I added the create callbacks to the tasking layer and
the tests I failed to update those, so they were too small and thus we
sometimes reported the callbacks or, for tcb-unin, uninstalled the first
one before all the expected callbacks had occurred.  This resulted in
intermittent failures.

While I was here I cleaned up a bit:
- give a "tcb_" prefix to all the C functions
- make the tcb_wait_for_nCallbacks() call explicit, rather than implicit
  in the uninstall and report calls
- remove the pass-through functions in tcb*-c.c in favor of calling the
  underlying common code directly.