More task callbacks test simplification.
Here we rearrange the code and rename some things so that all the C
stuff is in tcb-util.h and the .chpl code just refers to functions in
there.  We include this rather than compiling into a separate .o, so now
we don't need the .precomp or .cleanfiles files.  We also don't need the
-c.c files, and the -c.h files are renamed to just .h.  Now they just
define the size of the callback recording buffer and include tcb-util.h.