Remove specified return type from iterators in FileSystem.chpl
This is a workaround for the fact that normalization does not provide
initialization for the return value variable generated for functions with
specified return types.  Default-initialization is removed as an optimization,
under the assumption that the RVV must be assigned before the function exits.

However, assignment is not initialization, so in general the assignment will
fail.  It only happens to work for fundamental types because for those types
assignment and initialization are indistinguishable.