Make sed REs work properly on Mac.
GNU sed on Linux and sed on the Mac differ in terms of the differences
between basic (default) and extended regular expressions (REs).  In
particular, the '\|' and '\+' elements indicate alternatives and
one-or-more-instances in basic REs in GNU sed don't exist in Mac sed.
Here we simplify the REs so that they work on both.

We could instead put sed in extended-RE mode, but that would be much
more trouble because doing so requires a different command line option
for GNU sed and Mac sed.