Respect Chapel intents in extern proc declarations.
During development, I added code to treat a plain record type (default intent)
in extern proc declarations as "const in" -- a special case for records types.
After some discussion, we concluded it would be better to preserve the meaning
of Chapel intents and instead change extern proc declarations in the module
code.

This change was made on the string-as-rec branch and merged, so now the
special-case code can be removed from constIntentForType().