Add isRefIterType functionality to compiler and modules
Add a proc that returns whether something is a ref iterator for the purpose of
allowing us to create "wrapper" iterators. When an iterator is wrapping another
iterator we need to know the ref-ness of the wrapped iterator so we can specify
the refness of the wrapping iterator.

previously if we had something like:

  ``` chapel
  iter wrapIter(myIter) {
    writeln("Wrap!");
    for i in myIter do yield i;
  }

  // since wrapIter isn't a ref iter, we'll get an const-checking error
  for a in wrapIter(Arr) do a = 1;
  ```

now we can do someting like:

  ``` chapel
  iter wrapIter(myIter) ref where isRefIter(_getIterator(myIter)) {
    writeln("Wrap!");
    for i in myIter do yield i;
  }

  iter wrapIter(myIter) where !isRefIter(_getIterator(myIter)) {
    writeln("Wrap!");
    for i in myIter do yield i;
  }

  // will work as expected
  for a in wrapIter(Arr) do a = 1;
  // will correctly get const-checking error since val iter will be invoked
  for i in wrapIter(1..10) do i = 1;
  ```

This commit adds code to the modules and compiler modeled after isAtomicType().
It also adds a pretty extensive set of tests that checks if isRefIter works for
standalone iters, iters in record, iters in classes, and that isRefIter doesn't
break if called on a non-iterators.

This is to support adding a wrapping vectorize only iterator. Ideally the
signature of a wrapper would be something more like:

  ```chapel
  iter wrapIter(myIter: iterable) ref=isRefIter(getIterator(myIter)) {}
  ```