Add ability to explicitly mark iterators as vectorizable
Previously only yielding loops in standalone and follower iterators were marked
as order independent and possible considered as vectorization candidate. Now
iterators can be marked with a simple pragma to denote that their yielding
loops should also be marked as order independent.

This is to support a vectorize only wrapping iterator that will allow a serial
for loop to be explicitly vectorized. The pragma is not meant to be used by
users, but it will be applied to a wrapper iterator they can use. The patch
itself is really simple since it just adds another iterator type to examine for
order independent loops and the rest of the vectorization machinery just works
as is.