Fix errors with warning message insertion for invalid private declarations
Previous check to clamp down on private declarations in function bodies and
nested scopes caused multiple variable declarations to break.  This is because
multiple variable declarations insert a "scopeless" block around them, which is
a block statement with a block tag of BLOCK_SCOPELESS.  This seems a little
obnoxious, but perhaps compiler updates in the works will improve this
situation?  Anyways, with this change, the warning messages are applied
correctly.