Track zippered ForLoops throughout compilation
Previously the "zip-ness" of an iterator was lost during AST building. It was
used to build the ForLoop but there was no reliable way to determine if a
ForLoop was iterating over a zippered iterator after parsing.

This adds zippered as an argument to the ForLoop's constructor which is then
stored as a private immutable member variable that has an assessor method.

This will be used in an upcoming patch to prevent zippering of certain
iterators. It will be used to find zippered iterators and check for a pragma
that denotes that zippering illegal.

The zippered state was already passed to the factory builder `buildForLoop` so
we just forward that to the actual ForLoop constructor. The actual ForLoop
constructor is directly called from a few other builders so they just had to be
updated to take the appropriate zippered state.

In order of occurrence in build.cpp:

 - In buildFollowLoop, we pass the zippered state along to the follower
 - In buildStandaloneForallLoopStmt, standalone iters can't be zipped
 - In buildForallLoopStmt, we pass the zippered state along to the leader
 - In buildReduceExpr, we pass the zipperd state to the follower and leader

Note that as stated above, the ForLoops that invoke leader iterators will be
marked as zippered if the forall was zippered. You could argue that leaders
can't really be zippered since there can only ever be a single leader. However
something like:

```chapel
forall i in zip(myIter1(), myIter2()) do writeln(i);
```

is effectively transformed into:

```chapel
for i in zip(myIter(tag=iterKind.leader)) do
  for j in zip(myIter1(tag=iterKind.follower, followthis=i),
               myIter2(tag=iterKind.follower, followThis=i) do writeln(j);
```

and the ForLoop invoking the leader is still zippered even though there can
never be more than one leader.